pyproject.toml
[project]
name = "inst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "instructor>=1.6.3",
    "openai>=1.53.0",
    "pygithub>=2.4.0",
    "requests>=2.32.3",
    "supabase>=2.9.1",
]
